pyproject.toml
[tool.poetry]
name = "fast-weather"
version = "0.1.0"
description = ""
authors = ["Felipe Limeira <limeira.felipe94@gmail.com>"]
readme = "README.md"
packages = [{include = "fast_weather"}]

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
redis = "^5.0.7"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
